+++ tests/components/dhcp/test_init.py
@@ -3,7 +3,8 @@ import datetime
 import threading
 from unittest.mock import MagicMock, patch
 
-from scapy import arch  # pylint: unused-import  # noqa: F401
+import pytest
+from scapy import arch  # pylint: disable=unused-import  # noqa: F401
 from scapy.error import Scapy_Exception
 from scapy.layers.dhcp import DHCP
 from scapy.layers.l2 import Ether
@@ -845,6 +846,7 @@ async def test_aiodiscover_finds_new_hosts_after_interval(hass):
     )
 
 
+@pytest.mark.usefixtures("mock_integration_frame")
 async def test_service_info_compatibility(hass, caplog):
     """Test compatibility with old-style dict.
 
@@ -856,21 +858,13 @@ async def test_service_info_compatibility(hass, caplog):
         macaddress="b8b7f16db533",
     )
 
-    # Ensure first call get logged
-    assert discovery_info["ip"] == "192.168.210.56"
-    assert discovery_info.get("ip") == "192.168.210.56"
+    with patch("homeassistant.helpers.frame._REPORTED_INTEGRATIONS", set()):
+        assert discovery_info["ip"] == "192.168.210.56"
+    assert "Detected integration that accessed discovery_info['ip']" in caplog.text
+
+    with patch("homeassistant.helpers.frame._REPORTED_INTEGRATIONS", set()):
+        assert discovery_info.get("ip") == "192.168.210.56"
+    assert "Detected integration that accessed discovery_info.get('ip')" in caplog.text
+
     assert discovery_info.get("ip", "fallback_host") == "192.168.210.56"
     assert discovery_info.get("invalid_key", "fallback_host") == "fallback_host"
-    assert "Detected code that accessed discovery_info['ip']" in caplog.text
-    assert "Detected code that accessed discovery_info.get('ip')" not in caplog.text
-
-    # Ensure second call doesn't get logged
-    caplog.clear()
-    assert discovery_info["ip"] == "192.168.210.56"
-    assert discovery_info.get("ip") == "192.168.210.56"
-    assert "Detected code that accessed discovery_info['ip']" not in caplog.text
-    assert "Detected code that accessed discovery_info.get('ip')" not in caplog.text
-
-    discovery_info._warning_logged = False  # pylint: disable=[protected-access]
-    assert discovery_info.get("ip") == "192.168.210.56"
-    assert "Detected code that accessed discovery_info.get('ip')" in caplog.text
