+++ tests/components/zeroconf/test_init.py
@@ -3,6 +3,7 @@ from ipaddress import ip_address
 from typing import Any
 from unittest.mock import call, patch
 
+import pytest
 from zeroconf import InterfaceChoice, IPVersion, ServiceStateChange
 from zeroconf.asyncio import AsyncServiceInfo
 
@@ -1032,6 +1033,7 @@ async def test_no_name(hass, mock_async_zeroconf):
     assert info.name == "Home._home-assistant._tcp.local."
 
 
+@pytest.mark.usefixtures("mock_integration_frame")
 async def test_service_info_compatibility(hass, caplog):
     """Test compatibility with old-style dict.
 
@@ -1046,21 +1048,15 @@ async def test_service_info_compatibility(hass, caplog):
         properties={},
     )
 
-    # Ensure first call get logged
-    assert discovery_info["host"] == "mock_host"
-    assert discovery_info.get("host") == "mock_host"
+    with patch("homeassistant.helpers.frame._REPORTED_INTEGRATIONS", set()):
+        assert discovery_info["host"] == "mock_host"
+    assert "Detected integration that accessed discovery_info['host']" in caplog.text
+
+    with patch("homeassistant.helpers.frame._REPORTED_INTEGRATIONS", set()):
+        assert discovery_info.get("host") == "mock_host"
+    assert (
+        "Detected integration that accessed discovery_info.get('host')" in caplog.text
+    )
+
     assert discovery_info.get("host", "fallback_host") == "mock_host"
     assert discovery_info.get("invalid_key", "fallback_host") == "fallback_host"
-    assert "Detected code that accessed discovery_info['host']" in caplog.text
-    assert "Detected code that accessed discovery_info.get('host')" not in caplog.text
-
-    # Ensure second call doesn't get logged
-    caplog.clear()
-    assert discovery_info["host"] == "mock_host"
-    assert discovery_info.get("host") == "mock_host"
-    assert "Detected code that accessed discovery_info['host']" not in caplog.text
-    assert "Detected code that accessed discovery_info.get('host')" not in caplog.text
-
-    discovery_info._warning_logged = False
-    assert discovery_info.get("host") == "mock_host"
-    assert "Detected code that accessed discovery_info.get('host')" in caplog.text
